2D board
This commit allows the user to move a character following x and y axis,
the playing area is defined in a 100*x array. Terminal width must be
set to 100 columns in order for the game to work properly.

Commands are Z,Q,S,D
Exit key is ESC

The new code is included in keys.py
Original story in main
Other files are designs